src/neon_void/utils/sound_manager.py
"""
Sound effects manager for NEON VOID OPTIMIZER.
Cyberpunk-themed UI sounds (optional, toggleable).
"""

import logging
import os
import threading
import time
from typing import Optional

logger = logging.getLogger("NEON_VOID")


class SoundManager:
    """
    Manages optional cyberpunk sound effects for UI interactions.
    Falls back gracefully if audio libraries are not available.
    """

    SOUND_EFFECTS = {
        "hover": "short_electric_buzz",
        "click": "digital_click",
        "boost": "power_surge",
        "alert": "warning_chime",
        "success": "positive_chime",
        "startup": "system_boot",
    }

    def __init__(self) -> None:
        self.enabled = False
        self._sound_cache = {}
        self._initialized = False

    def initialize(self) -> bool:
        """Initialize the sound system."""
        if self._initialized:
            return True

        try:
            import sounddevice as sd
            import soundfile as sf
            self._sd = sd
            self._sf = sf
            self._initialized = True
            logger.info("Sound system initialized")
            return True

        except ImportError:
            logger.debug("sounddevice/soundfile not available")
            return False

    def play(self, sound_name: str) -> None:
        """Play a named sound effect."""
        if not self.enabled or not self._initialized:
            return

        def _play():
            try:
                # In production, load actual sound files
                # For now, this is a placeholder
                pass
            except Exception as e:
                logger.debug(f"Sound play error: {e}")

        # Play in background thread to avoid blocking UI
        thread = threading.Thread(target=_play, daemon=True)
        thread.start()

    def play_hover(self) -> None:
        """Play hover sound effect."""
        self.play("hover")

    def play_click(self) -> None:
        """Play click sound effect."""
        self.play("click")

    def play_boost(self) -> None:
        """Play boost activation sound."""
        self.play("boost")

    def play_success(self) -> None:
        """Play success sound."""
        self.play("success")

    def play_alert(self) -> None:
        """Play alert sound."""
        self.play("alert")

    def set_enabled(self, enabled: bool) -> None:
        """Enable or disable sound effects."""
        self.enabled = enabled
        if enabled and not self._initialized:
            self.initialize()
        logger.info(f"Sound effects {'enabled' if enabled else 'disabled'}")


# Global instance
sound_manager = SoundManager()
